+++ src/Dilcore.MongoDB/Extensions/ServiceCollectionExtensions.cs
@@ -1,5 +1,6 @@
 using System.Linq.Expressions;
 using Dilcore.MongoDB.Abstractions;
+using Dilcore.MongoDB.Abstractions.Json;
 using Dilcore.MongoDB.Abstractions.Keys;
 using Dilcore.MongoDB.Abstractions.Namespace;
 using Dilcore.MongoDB.Abstractions.Options;
@@ -8,6 +9,7 @@ using Dilcore.MongoDB.Abstractions.Transactions;
 using Dilcore.MongoDB.DependencyInjection;
 using Dilcore.MongoDB.Descriptors;
 using Dilcore.MongoDB.Internal;
+using Dilcore.MongoDB.Json;
 using Dilcore.MongoDB.Namespace;
 using Dilcore.MongoDB.Repositories;
 using Dilcore.MongoDB.Transactions;
@@ -44,6 +46,7 @@ public static class ServiceCollectionExtensions
                 sp.GetRequiredService<MongoRegistrationGraph>()));
         services.AddScoped<IMongoDatabaseResolver, MongoDatabaseResolver>();
         services.AddScoped<IMongoDbCollectionFactory, MongoDbCollectionFactory>();
+        services.AddSingleton<IBsonJsonConverter, BsonJsonConverter>();
         services.AddScoped<IMongoDbTransactionRunner, MongoDbTransactionRunner>();
         services.AddScoped<IRepositoryResolver, RepositoryResolver>();